poll-limit: fix warning

+++ activitypub.c
@@ -17,6 +17,7 @@
 
 #include "snac.h"
 
+#include <stddef.h>
 #include <sys/wait.h>
 
 const char * const public_address = "https:/" "/www.w3.org/ns/activitystreams#Public";
@@ -2321,8 +2322,8 @@ xs_dict *msg_question(snac *user, const char *content, xs_list *attach,
 /* creates a Question message */
 {
     xs_dict *msg = msg_note(user, content, NULL, NULL, attach, 0, NULL, NULL);
-    int max_line = 200;
-    int max      = 8;
+    size_t max_line = 200;
+    int max = 8;
     const xs_number *max_options = xs_dict_get(srv_config, "max_poll_options");
     xs_set seen;
